# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: TKDL – Tesco Kilbirnie Darts League API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches (optionally filtered by season)
      parameters:
        - name: seasonId
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: submitMatch
      tags: [matches]
      summary: Submit a match result with wager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Match recorded, standings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "400":
          description: Validation error
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get a match by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Not found
    delete:
      operationId: deleteMatch
      tags: [matches]
      summary: Delete / void a match result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Seasons ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Match:
      type: object
      required: [id, seasonId, winnerId, loserId, winnerName, loserName, stake, eloChange, playedAt]
      properties:
        id:
          type: integer
        seasonId:
          type: integer
        winnerId:
          type: integer
        loserId:
          type: integer
        winnerName:
          type: string
        loserName:
          type: string
        stake:
          type: integer
          description: Points wagered on this match
        eloChange:
          type: integer
        gameType:
          type: string
        isTeamMatch:
          type: boolean
        notes:
          type: ["string", "null"]
        playedAt:
          type: string
          format: date-time
        winnerDarts:
          type: integer
        winner100s:
          type: integer
        winner140s:
          type: integer
        winner170s:
          type: integer
        winner180s:
          type: integer
        winnerCheckoutAttempts:
          type: integer
        winnerCheckoutHits:
          type: integer
        loserDarts:
          type: integer
        loser100s:
          type: integer
        loser140s:
          type: integer
        loser170s:
          type: integer
        loser180s:
          type: integer
        loserCheckoutAttempts:
          type: integer
        loserCheckoutHits:
          type: integer
    MatchInput:
      type: object
      required: [winnerId, loserId, stake]
      properties:
        winnerId:
          type: integer
        loserId:
          type: integer
        stake:
          type: integer
          minimum: 0
          description: Points wagered (max = min of both balances)
        gameType:
          type: string
        notes:
          type: string
        winnerDarts:
          type: integer
        winner100s:
          type: integer
        winner140s:
          type: integer
        winner170s:
          type: integer
        winner180s:
          type: integer
        winnerCheckoutAttempts:
          type: integer
        winnerCheckoutHits:
          type: integer
        loserDarts:
          type: integer
        loser100s:
          type: integer
        loser140s:
          type: integer
        loser170s:
          type: integer
        loser180s:
          type: integer
        loserCheckoutAttempts:
          type: integer
        loserCheckoutHits:
          type: integer

    # ── Season ─────────────────────────────────────────────────────────────────
...